Struct unqlite::UnQLite

pub struct UnQLite { /* fields omitted */ }

UnQLite database entry point.

UnQLite support both in-memory and on-disk database. There's several constructors:

Constructor 	 Meaning

create_in_memory 	 Create a private, in-memory database.

create_temp 	 Create a private, temporary on-disk database.

create 	 Create if not exists, otherwise, open as read-write.

open_mmap 	 Obtain a read-only memory view of the whole database.

open_readonly 	 Open the database in a read-only mode.

Methods

impl UnQLite

pub fn create<P: AsRef<str>>(filename: P) -> UnQLite

Create UnQLite database as filename.

By default, the database is created in read-write mode.

Panics

Will panic if failed in creating.

Example

let _ = UnQLite::create("test.db");

C

unqlite *pDb;

// on-disk database
rc = unqlite_open(&pDb,"test.db",UNQLITE_OPEN_CREATE);

// in-memory database
rc = unqlite_open(&pDb, ":mem:", UNQLITE_OPEN_MEM);

pub fn create_in_memory() -> UnQLite

Create database in memory.

Equivalent to:

let _ = UnQLite::create(":mem:");

Panics

Will panic if failed in creating.

pub fn create_temp() -> UnQLite

A private, temporary on-disk database will be created.

This private database will be automatically deleted as soon as the database connection is closed.

Panics

Will panic if failed in creating.

C

int rc = unqlite_open("test.db", UNQLITE_OPEN_TEMP_DB);

pub fn open_mmap<P: AsRef<str>>(filename: P) -> UnQLite

Obtain a read-only memory view of the whole database.

You will get significant performance improvements with this combination but your database is still read-only.

Panics

Panic if open failed.

C

unqlite_open(&pDb, "test.db", UNQLITE_OPEN_MMAP | UNQLITE_OPEN_READONLY);

pub fn open_readonly<P: AsRef<str>>(filename: P) -> UnQLite

Open the database in a read-only mode.

That is, you cannot perform a store, append, commit or rollback operations with this control flag.

Always prefer to use open_mmap for readonly in disk database.

Panics

Panic too.

C

unqlite_open(&pDb, "test.db", UNQLITE_OPEN_READONLY);